Moved Popup and BasePopup to their own class files
